Ask user if they want to install new toolchains
Also add some user output for the major steps of installing toolchains,
compiling `spirv-builder-cli` and compiling the actual shader.

Surprisingly there's no way in the standard library to get a single
keypress from the user! There's ways to get a line, therefore a keypress
then a newline, but that's not so conventional for responding to a "y/n"
prompt.

Fixes #14